[BUGFIX] Change sys_file.size type from INT to BIGINT
Regular INT limits filesize to 2GB.

Resolves: #60327
Releases: 6.3, 6.2
Reviewed-on: http://review.typo3.org/31737
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Susanne Moog <typo3@susannemoog.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>

+++ typo3/sysext/core/ext_tables.sql
@@ -302,7 +302,7 @@ CREATE TABLE sys_file (
 	mime_type varchar(255) DEFAULT '' NOT NULL,
 	name tinytext,
 	sha1 tinytext,
-	size int(11) DEFAULT '0' NOT NULL,
+	size bigint(20) unsigned DEFAULT '0' NOT NULL,
 	creation_date int(11) DEFAULT '0' NOT NULL,
 	modification_date int(11) DEFAULT '0' NOT NULL,
